Solution to problem 25
I was getting tired of using vectors to store and operate on very large
integers, so I found a nice lightweight big int library.

Also, I was having some problems with MinGW, so I switched over to
MinGW-w64. So far it seems that everything that wasn't working with
MinGW is working now.